Precompile regexes used to check invalid double star
The match() function is expected to be called many times, and
using re.search() means the regexes used to check invalid **
usage will be recompiled on every run. Even though this is
unlikely to be a bottleneck, it's better to compile the regexes
once and then use the compiled versions.